extract plotting utils
